Feat: 예약하기 구현

+++ BE/src/main/java/com/team03/airbnb/controller/SearchController.java
@@ -33,7 +33,7 @@ public class SearchController {
     }
 
     @GetMapping("/detail/{id}")
-    public HotelDetailDTO findHotelDetailById(@PathVariable Integer id) {
+    public HotelDetailDTO findHotelDetailById(@PathVariable Long id) {
         HotelDetailDTO hotelDetailDTO = hotelService.findHotelDetailById(id);
         return hotelDetailDTO;
     }
